Moves "isStringEqual" to internal utilities
