+++ src/utils/internal/isStringEqual.test.ts
